asm-verify: fix ulong/int width + const signature mismatches (#266 Cluster D)
Restore binary signatures so the mangled names pair (unsigned long = m,
unsigned int = j, int = i -- distinct even where all 4 bytes on the port):
- Font::GetPage(unsigned long) const -- was int
- Skeleton::BuildArrays(unsigned long) -- was int
- Touch::FindTouch/GetTouchPos/GetTouchDelta -- param was already uint;
  real divergence was const-ness (binary is non-const) -> drop const
- DojoScreen::UpdateBSButton(BSButton*, float, unsigned long) -- 3rd was int
- Bonus::IsAchieved map key std::map<unsigned long,int> -- was uint32_t
  (j); changed m_MinFruit/m_MaxFruit/m_RequiredHashes keys + IsAchieved/
  GetBest/UnlockAchievements iterators consistently. Behavior-safe:
  unsigned long is 4 bytes on ARM32/Windows LLP64, sizeof(std::map)
  unchanged, offset static_asserts still hold, hashes are 32-bit anyway.

Mangling-only on the port (LLP64), but fixes asm-verify pairing on the
Bada cross-build. Call sites pass int literals -> implicit convert, no
call-site changes. triage.json was partly stale (Touch param + Bonus key
were not what it claimed) -- verified each vs current source first.

Verified: fresh build + 79/79 logic ctest.

+++ src/game/Bonus.cpp
@@ -213,7 +213,7 @@ void Bonus::Parse(TiXmlElement* e) {
 //   5. return m_Tier (0 on any fail).
 // DIFFERS: original param name was `score` -- renamed to `score` kept; the
 //   second param is fruitCounts (binary param_2). Cosmetic only; no ABI change.
-int Bonus::IsAchieved(int score, std::map<uint32_t, int>& fruitCounts) {
+int Bonus::IsAchieved(int score, std::map<unsigned long, int>& fruitCounts) {
     // Gate -- unconditional bounds + divisible-by (binary 0010df3e-0010df62).
     if (score < m_MinSliced) return 0;
     if (score > m_MaxSliced) return 0;
@@ -223,14 +223,14 @@ int Bonus::IsAchieved(int score, std::map<uint32_t, int>& fruitCounts) {
     // Missing key in m_MinFruit defaults to 0; missing key in m_MaxFruit defaults
     // to DAT_0010e090 = 1,000,000.
     static const int kNoMaxSentinel = 1000000;  // DAT_0010e090 = 0x000f4240
-    for (std::map<uint32_t, int>::iterator fc = fruitCounts.begin();
+    for (std::map<unsigned long, int>::iterator fc = fruitCounts.begin();
          fc != fruitCounts.end(); ++fc) {
         int count = fc->second;
 
-        std::map<uint32_t, int>::const_iterator minIt = m_MinFruit.find(fc->first);
+        std::map<unsigned long, int>::const_iterator minIt = m_MinFruit.find(fc->first);
         int minVal = (minIt != m_MinFruit.end()) ? minIt->second : 0;
 
-        std::map<uint32_t, int>::const_iterator maxIt = m_MaxFruit.find(fc->first);
+        std::map<unsigned long, int>::const_iterator maxIt = m_MaxFruit.find(fc->first);
         int maxVal = (maxIt != m_MaxFruit.end()) ? maxIt->second : kNoMaxSentinel;
 
         if (count < minVal || count > maxVal) return 0;
@@ -241,7 +241,7 @@ int Bonus::IsAchieved(int score, std::map<uint32_t, int>& fruitCounts) {
     bool firstPattern = true;
     int refCount = -1;
     for (size_t i = 0; i < m_PatternHashes.size(); ++i) {
-        std::map<uint32_t, int>::iterator fc = fruitCounts.find(m_PatternHashes[i]);
+        std::map<unsigned long, int>::iterator fc = fruitCounts.find(m_PatternHashes[i]);
         if (fc == fruitCounts.end()) return 0;  // pattern fruit absent
         if (firstPattern) {
             refCount = fc->second;
@@ -352,9 +352,9 @@ int GetBonusTotal(unsigned long hash);
 // ASM-verified: 2026-05-22 v1.6.1 BonusType::GetBest @ 0x0010ecc8 (re-analyst).
 Bonus* BonusType::GetBest() {
     int totalAcrossFruits = 0;
-    for (std::map<uint32_t, int>::iterator it = m_RequiredHashes.begin();
+    for (std::map<unsigned long, int>::iterator it = m_RequiredHashes.begin();
          it != m_RequiredHashes.end(); ++it) {
-        int total = GetBonusTotal((unsigned long)it->first);
+        int total = GetBonusTotal(it->first);
         it->second = total;
         totalAcrossFruits += total;
     }
@@ -404,9 +404,9 @@ bool BonusType::UnlockAchievements() {
     if (!m_HasAchievement) return false;
 
     int totalAcrossFruits = 0;
-    for (std::map<uint32_t, int>::iterator it = m_RequiredHashes.begin();
+    for (std::map<unsigned long, int>::iterator it = m_RequiredHashes.begin();
          it != m_RequiredHashes.end(); ++it) {
-        int total = GetBonusTotal((unsigned long)it->first);
+        int total = GetBonusTotal(it->first);
         it->second = total;
         totalAcrossFruits += total;
     }
